+++ connect.go
@@ -9,7 +9,7 @@ import (
 
 var tick int32
 
-func dial(network string, hosts []string, noDelay bool, r, w time.Duration, log func(string, ...interface{})) (*connect, error) {
+func dial(network string, hosts []string, noDelay bool, r, w time.Duration, logf func(string, ...interface{})) (*connect, error) {
 	var (
 		err error
 		abs = func(v int) int {
@@ -23,13 +23,13 @@ func dial(network string, hosts []string, noDelay bool, r, w time.Duration, log 
 	)
 	for i := 0; i <= len(hosts); i++ {
 		if conn, err = net.DialTimeout(network, hosts[(index+1)%len(hosts)], 2*time.Second); err == nil {
-			log("[connect] num=%d -> %s", tick, conn.RemoteAddr())
+			logf("[connect] num=%d -> %s", tick, conn.RemoteAddr())
 			if tcp, ok := conn.(*net.TCPConn); ok {
 				tcp.SetNoDelay(noDelay) // Disable or enable the Nagle Algorithm for this tcp socket
 			}
 			return &connect{
 				Conn:         conn,
-				log:          log,
+				logf:         logf,
 				readTimeout:  r,
 				writeTimeout: w,
 			}, nil
@@ -40,7 +40,7 @@ func dial(network string, hosts []string, noDelay bool, r, w time.Duration, log 
 
 type connect struct {
 	net.Conn
-	log          func(string, ...interface{})
+	logf         func(string, ...interface{})
 	readTimeout  time.Duration
 	writeTimeout time.Duration
 }
@@ -51,7 +51,7 @@ func (conn *connect) Read(b []byte) (int, error) {
 	}
 	n, err := conn.Conn.Read(b)
 	if err != nil {
-		conn.log("[connect] read error: %v", err)
+		conn.logf("[connect] read error: %v", err)
 		return n, driver.ErrBadConn
 	}
 	return n, nil
@@ -63,7 +63,7 @@ func (conn *connect) Write(b []byte) (int, error) {
 	}
 	n, err := conn.Conn.Write(b)
 	if err != nil {
-		conn.log("[connect] write error: %v", err)
+		conn.logf("[connect] write error: %v", err)
 		return n, driver.ErrBadConn
 	}
 	return n, nil
